% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censuspyrID-package.R
\docType{package}
\name{censuspyrID-package}
\alias{censuspyrID}
\alias{censuspyrID-package}
\title{censuspyrID: Explore and visualize Indonesian census pyramids}
\description{
Provides harmonized and unharmonized population pyramid datasets from the Indonesian population censuses (1971–2020). The package includes tools for visualization and an interactive Shiny application for data exploration.
}
\author{
\strong{Maintainer}: Ari Purwanto Sarwo Prasojo \email{ari.prasojo18@gmail.com} (\href{https://orcid.org/0000-0002-4862-5523}{ORCID}) (arip003@brin.go.id)

Authors:
\itemize{
  \item Puguh Prasetyoputra \email{pprasetyoputra@gmail.com} (\href{https://orcid.org/0000-0001-5494-7003}{ORCID}) (pugu004@brin.go.id)
  \item Nur Fitri Mustika Ayu \email{nurfitrimustikaayu@gmail.com}
}

}
\keyword{internal}
