% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censuspyrID.R
\name{censuspyrID_explorer}
\alias{censuspyrID_explorer}
\title{Explore Harmonized and Unharmonized Population Pyramids from Indonesia’s Censuses (1971–2020)}
\usage{
censuspyrID_explorer(host = NULL, ...)
}
\arguments{
\item{host}{Character string passed to \code{\link[shiny:runApp]{runApp}}.
Default is \code{"0.0.0.0"}.}

\item{...}{Additional arguments passed to \code{\link[shiny:runApp]{runApp}}.}
}
\value{
The function launches the Shiny application. It does not return a value.
}
\description{
Launches \strong{censuspyrID Explorer}, a Shiny application for visualizing harmonized and
unharmonized population pyramids from Indonesia’s population censuses (1971–2020).
}
\details{
The application provides interactive tools to explore demographic structures across provinces
and census years. See the \emph{Help} menu within the application for a navigation guide.
}
\examples{
\dontrun{
censuspyrID_explorer()
}

}
