% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/censuspyrID.R
\name{reg_list}
\alias{reg_list}
\title{Create region (province) list}
\usage{
reg_list(harmonized = TRUE)
}
\arguments{
\item{harmonized}{Logical. If \code{TRUE} (default), use harmonized province codes.
If \code{FALSE}, use unharmonized province codes.}
}
\value{
A named vector, where values are province IDs and names are
corresponding province labels.
}
\description{
Internal helper function to generate a list of provinces,
depending on whether harmonized or unharmonized coding is used.
}
\details{
\itemize{
\item If \code{harmonized = TRUE}, the function uses \code{ref_label$provinceh_label}.
\item If \code{harmonized = FALSE}, the function uses \code{ref_label$province_label}.
}
}
\keyword{internal}
